# openapi.yaml
openapi: 3.0.0
info:
  title: Service Envelope API v2.0
  version: 2.1.0
  description: REST API for the Service Envelope System. Includes OTP bearer
    authentication, Mongo-backed auth users and roles, generic email template
    fallbacks, schema v1.0.5 service validation, and unified delivery selection.
  contact:
    name: Service Envelope Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development
  - url: https://api.mapua.edu.ph
    description: Production
paths:
...
  /api/requests:
    get:
      tags:
        - Requests
      summary: List All Requests
      operationId: listAllRequests
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusQueryParam"
        - $ref: "#/components/parameters/TypeQueryParam"
        - $ref: "#/components/parameters/LimitQueryParam"
        - $ref: "#/components/parameters/OffsetQueryParam"
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    total: 1
                    count: 1
                    limit: 1
                    offset: 1
                    requests:
                      - id: sample
                        type: sample
                        initiator: sample
                        status: sample
                        createdAt: 2026-06-07T00:00:00.000Z
                        lastUpdated: 2026-06-07T00:00:00.000Z
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET "http://localhost:8000/api/requests"
    post:
      tags:
        - Requests
      summary: Create Request
      operationId: createRequest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCreatePayload"
            examples:
              usingServiceId:
                value:
                  serviceId: SERVICE-999
                  initiator: requester@mapua.edu.ph
                  parameters:
                    studentId: 20250001
                    firstName: Juan
                    lastName: Dela Cruz
                    email: juan.delacruz@mapua.edu.ph
              usingType:
                value:
                  type: comprehensive-student-document
                  initiator: requester@mapua.edu.ph
                  parameters:
                    studentId: 20250001
                    firstName: Juan
                    lastName: Dela Cruz
                    email: juan.delacruz@mapua.edu.ph
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRequestResponse"
              examples:
                sample:
                  summary: Sample 201 response
                  value:
                    success: true
                    requestId: REQ-20260606-042
                    type: sample
                    serviceId: sample
                    status: queued
                    envelopes:
                      sampleKey: sampleValue
                    message: sample
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X POST "http://localhost:8000/api/requests" \
              -H "Content-Type: application/json" \
              -d '{"sample":true}'
  /api/requests/{requestId}:
    get:
      tags:
        - Requests
      summary: Get Request Details
      operationId: getRequestDetails
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetailsResponse"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    id: sample
                    type: sample
                    initiator: sample
                    status: sample
                    createdAt: 2026-06-07T00:00:00.000Z
                    lastUpdated: 2026-06-07T00:00:00.000Z
                    envelopes:
                      sampleKey: sampleValue
                    history:
                      - sampleKey: sampleValue
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET "http://localhost:8000/api/requests/REQ-20260606-865"
    delete:
      tags:
        - Requests
      summary: Cancel Request
      operationId: cancelRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Request cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                    example: REQ-20260606-042
                  status:
                    type: string
                    example: cancelled
                  message:
                    type: string
                    example: Request cancelled successfully
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: REQ-20260606-042
                    status: cancelled
                    message: Request cancelled successfully
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X DELETE "http://localhost:8000/api/requests/REQ-20260606-865"
  /api/requests/{requestId}/resume:
    post:
      tags:
        - Requests
      summary: Resume Request
      operationId: resumeRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
            examples:
              sample:
                summary: Sample request
                value: {}
      responses:
        "200":
          description: Request resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  message:
                    type: string
                  status:
                    type: string
                    example: processing
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: sample
                    message: sample
                    status: processing
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: >-
            curl -X POST
            "http://localhost:8000/api/requests/REQ-20260606-865/resume" \
              -H "Content-Type: application/json" \
              -d '{}'
  /api/requests/{requestId}/history:
    get:
      tags:
        - Requests
      summary: Get Request History
      operationId: getRequestHistory
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Request history
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  history:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  totalEntries:
                    type: integer
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    requestId: sample
                    history:
                      - sampleKey: sampleValue
                    totalEntries: 1
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/requests/REQ-20260606-865/history"
...
  /api/approvals/{token}/request:
    get:
      tags:
        - Approvals
      summary: Get Request Details for Approval Token
      operationId: getApprovalTokenRequest
      security: []
      parameters:
        - $ref: "#/components/parameters/TokenParam"
      responses:
        "200":
          description: Request details for approval UI
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    sampleKey: sampleValue
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/approvals/94c2fea5-7de6-427c-9da8-ad3948673177/request"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequestIdParam:
      name: requestId
      in: path
      required: true
      schema:
        type: string
      example: REQ-20260606-042
      description: Request identifier.
    TokenParam:
      name: token
      in: path
      required: true
      schema:
        type: string
      example: 2d2f5e0a-5a8f-4b7d-a365-6dd77c9f17d8
      description: Approval or feedback token.
...
    StatusQueryParam:
      name: status
      in: query
      required: false
      schema:
        type: string
      example: pending_payment
      description: Filter by request overall status.
    TypeQueryParam:
      name: type
      in: query
      required: false
      schema:
        type: string
      example: comprehensive-student-document
      description: Filter by request type.
    LimitQueryParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 1000
      example: 20
    OffsetQueryParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
...
  schemas:
...
    RequestCreatePayload:
      type: object
      required:
        - initiator
        - parameters
      properties:
        type:
          type: string
          description: Service type identifier.
          example: comprehensive-student-document
        serviceId:
          type: string
          description: Service ID identifier.
          example: SERVICE-999
        initiator:
          type: string
          description: Actor creating the request.
          example: requester@mapua.edu.ph
        parameters:
          type: object
          additionalProperties: true
      anyOf:
        - required:
            - type
        - required:
            - serviceId
    CreateRequestResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          example: REQ-20260606-042
        type:
          type: string
        serviceId:
          type: string
        status:
          type: string
          example: queued
        envelopes:
          type: object
          additionalProperties: true
        message:
          type: string
    RequestListResponse:
      type: object
      properties:
        total:
          type: integer
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        requests:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              initiator:
                type: string
              status:
                type: string
              createdAt:
                type: string
                format: date-time
              lastUpdated:
                type: string
                format: date-time
    RequestDetailsResponse:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        initiator:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        envelopes:
          type: object
          additionalProperties: true
        history:
          type: array
          items:
            type: object
            additionalProperties: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Request not found
        details:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 400 response
              value:
                error: Request not found
                details: sample
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 404 response
              value:
                error: Request not found
                details: sample
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 500 response
              value:
                error: Request not found
                details: sample
...